Issue #3593, #1197- fix isolate management
Please consider this change as a commit in the branch "issue-1197". In new v8 v8::Isolate does not get  instantiated automatically (inside `v8::Isolate::GetCurrent()`) anymore, so we have to manually call the creator method (`v8::Isolate::New()`) as well as it addresses the issue to free it when we don't need it.

The proposed change does work with current version of v8 as well as with a newer one.

The hack with `createJsEngine` is required for both versions.
It seems current version of v8 does not permit our approach in tests (create new Isolate in each JsEngine instance). I would suppose it's because the previous isolate "stored" in TLS is still "alive" but we are creating a new one.
Newer v8 is trying to avoid using of `v8::Isolate::GetCurrent` but despite of it, tests are still failing because starting from some point v8 cannot obtain more memory.

I think it's mainly because calls of setTimeout create background threads which keep JS engine from destroying and we consume all 2GB (available by default on 32bit Windows).