Issue #3594 - Remove `globalJsObject` and add `GetGlobalObject()`
Remove the global object member variable, which is part of the reference cycle
that is the subject of #3594. Replace it with a function that recreates an
equivalent JS variable each time. Change the single use of `globalJsObject` to
call `GetGlobalObject()` instead.

Added member variable and accessor for the v8 context in `JsContext`.