Issue 5198 - workaround for race condition in tests
It's merely a workaround to have tests running until it's properly solved.

The issue currently is that adding of a subscription triggers timeouts with zero seconds timeout. As the result when the test finishes JsEngine is not referenced anymore but the callback for zero timeout is being still executed and tries to destroy JsEngine and we hit the throwing of an exception from destructor (std::thread::join does it to avoid dead-lock). The workaround gives some time to finish these timer callbacks. So far, having such additional sleep only in that place is enough to run tests.