Issue 6526 - pass v8::Isolate to more functions because old approach is deprecated
Beside v8::Isolate some places now start to pass current context, though the implementation is not complete yet because there is no check of the return value.

Despite it can seem that static JsError::ExceptionToString should return an optinal/maybe value if there is no value, IMO it would be better to get as much as possible information, even just a string "unknown line" when it is about an exception. Perhaps even such information can already help to find out the bug, notice it or debug it.