Noissue - remove lib/publicSuffixList.js
The equivalent of this file is currently available from adblockpluscore at data/publicSuffixList.json. `require` and convert_js.py have been already adapted in earlier commits. In this commit, all mentioning of this file are removed as well as lib/basedomain.js is adapted. In the latter the global variable `publicSuffixes` is restored for the sake of safety, because it was a global var in the past and lib/basedomain.js is executed in the global context too. It seems that only this file is using it, though.