Noissue - keep the behavior of FilterEngine::Matches backward compatible
As it's mentioned in a previous commit of this series

The matcher returns a whitelist filter only if a blocking filter is found
 (issue 7003 [6]); there is a new CombinedMatcher#isWhitelisted() method to
 look up whitelist filters directly (issue 7162 [7]).

This is a breaking change. We should try to keep the behavior of C++ FilterEngine::Matches as well as methods based on it for some time for the sake of backward compatibility, and meanwhile add corresponding new methods and mark the former one as deprecated.

Pay attention that the implementation firstly returns an exception filter if there is one and only otherwise a blocking filter.

PS. follow up work is in DP-1088.