DP-1605: Make JsWeakValues really weak.
JsWeakValues makes no JsValue must be bound to the callback
passed to the embedder so it allows to preserve destruction order
because they are destructed with JsEngine. Otherwise we may end up
having JsValues destructed after JsEngine which is wrong. This was
actually the case in default file system implementation which wrapped
these in another structure which was passed to callbacks as a shared ptr
making them in fact to own the values.