DP-1943 Update Makefile and CI scripts to get the later v8 binaries
The Makefile and CI scripts have been re-arranged to get the latest prebuilt
binaries and to build under the new NDK. The later pre-built binaries have
their name changed compared to the old one. Also, the 7z is not used anymore
and the NDK has been updated to r20b. The commit reflects those updates.

Important to note that with this patch there is only one place that needs to be
updated in order to fetch the new prebuilt binaries when they are available.
This is the `V8_COMMIT` variable of Makefile and will affect both the local
builds and the CI builds.