DP-1947: Workaround for no control over destruction order in Java.
DP-1347 made ownership clear in libabp. Previosuly we had all shared
and despite that we still had hidden crahes due to cross dependencies.
This is now straightem up and the requirement is no JsValue can outlive
JsEngine. The problem is on Java (libabp-android) side where no
destruction order can be assured. It's also due to how libabp-android's
JNI & data structures are implemented and hopefully it'll also be
straighten up with DP-1940 and DP-1941 and we'll be able to get rid of
this workaround - but for time being in order to allow to upgrade without
having aforementioned tasks done let's have it - build time configurable.