DP-1943 Workaround to solve intermittent test case failure
The issue has been raised in core [1]. Utill, the issue is fixed that
workaround solves it. This commit can be reverted when the core is updated and
issue [1] is included.

The problem is that the `subscription.lastDownload` in `lib\init.js` is not
updated fast enough when a download is triggered. As a result, the same entry is
re-downloaded. It only happens with the latest V8 binaries

[1] https://gitlab.com/eyeo/adblockplus/adblockpluscore/-/issues/285