DP-2074: Keep one true global context owned by the engine.
Having multiple global contexts should be ok but we create them copying
from the one owned by engine just to work around not having possibility
to copy the global context (as opposed to creating it via New()).
Due to the fact it can be copied many global copies of it should be ok
but somehow this makes v8 unhappy and introduces weird dependencies between
contexts which may lead to random crashes.