DP-2074 Applied two patches that fix crash.
Patch 0

Having multiple global contexts should be ok but we create them copying
from the one owned by engine just to work around not having possibility
to copy the global context (as opposed to creating it via New()).
Due to the fact it can be copied many global copies of it should be ok
but somehow this makes v8 unhappy and introduces weird dependencies between
contexts which may lead to random crashes.

No handles should be accessed before locking of the isolate
if the isolate can be accessed from multiple threads.

Patch 1

Difference between GetJsValues/TakeJsValues sometimes unclear, so we
have several cases where we do just get but not take. It is not a strict
memory leak, but we keep some referred objects until engine is destroyed.