DPD-202: Report std deviation and std error in Harness Test results
Reporting min and max doesn't seem helpful because there are big outliers.
The outliers don't necessarily represent the cost of calling Core, rather
they seem to be an artifact of measurement in a noisy environment.

Instead, compute the standard deviation and standard error for the results.

https://towardsdatascience.com/standard-deviation-vs-standard-error-5210e3bc9c04:
"If you need to draw conclusions about the spread and variability of the data,
standard deviation is what you’ll need to use.
If you’re interested in finding how precise the sample mean is or you’re testing
the differences between two means, standard error is your metric."